Cambios

+++ database/migrations/2020_10_13_154801_create_consultorios_table.php
@@ -20,7 +20,7 @@ class CreateConsultoriosTable extends Migration
 
             $table->string('direccion',200);
             
-            $table->unsignedBigInteger('localidad_id'); 
+            $table->integer('localidad_id'); 
             $table->foreign('localidad_id')->references('id')->on('localidades');
         });
     }
